fix: fix coverage
