+++ jest.config.cjs
@@ -6,7 +6,7 @@ module.exports = {
   transform: {
     '^.+\\.js$': './config/babelJestTransform.cjs',
     '^.+\\.svelte$': ['svelte-jester', {
-      preprocess: './svelte.config.cjs',
+      preprocess: './config/svelte.config.cjs',
       compilerOptions: {
         dev: false
       }
@@ -34,8 +34,8 @@ module.exports = {
       lines: 100
     },
     './src/picker/components/Picker/Picker.svelte': {
-      statements: 98,
-      branches: 95,
+      statements: 96,
+      branches: 94,
       functions: 100,
       lines: 100
     }
